TS-29873 Add tests
